NewFormatOfRobot/WebsiteTesting/WebCourseLibrary.py
# File: WebCourseLibrary.py
from robot.api.deco import keyword
from code_validator import WebCourseValidator, DOMComparator
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import (
    TimeoutException,
    NoSuchElementException,
    WebDriverException,
)
import os
import json
import tempfile
import time


class WebCourseLibrary:

    ROBOT_LIBRARY_SCOPE = "GLOBAL"

    def __init__(self, timeout=30):
        self.validator = WebCourseValidator()
        self.last_result = None
        self.driver = None
        self.timeout = timeout
        self.wait = None

    @keyword
    def validate_html_structure(self, user_html, reference_html):
        """Validates HTML structure between user and reference HTML"""
        comparator = DOMComparator()
        result = comparator.compare_html_structure(user_html, reference_html)
        self.last_result = result
        return result["pass"]

    @keyword
    def validate_from_html_files(self, user_html_file, reference_html_file):
        """Validates HTML structure from files"""
        result = self.validator.validate_from_files(
            user_html_file=user_html_file, reference_html_file=reference_html_file
        )
        self.last_result = result
        return result["success"]

    @keyword
    def validate_complete_submission(
        self,
        user_html,
        user_css,
        user_js,
        reference_html,
        reference_css,
        reference_js,
        force_pass=False,
    ):
        """Validates complete submission with option to force pass"""
        result = self.validator.validate_submission(
            user_html, user_css, user_js, reference_html, reference_css, reference_js
        )

        # Force pass if requested
        if force_pass:
            result["success"] = True
            result["override_pass"] = True

        self.last_result = result
        return result["success"]

    @keyword
    def validate_from_files(
        self,
        user_html_file="",
        user_css_file="",
        user_js_file="",
        reference_html_file="",
        reference_css_file="",
        reference_js_file="",
        force_pass=False,
    ):
        """Validates submission from files with force pass option"""
        result = self.validator.validate_from_files(
            user_html_file,
            user_css_file,
            user_js_file,
            reference_html_file,
            reference_css_file,
            reference_js_file,
        )

        # Force pass if requested
        if force_pass:
            result["success"] = True
            result["override_pass"] = True

        self.last_result = result
        return result["success"]

    @keyword
    def validate_code(
        self, user_code, reference_code, language="html", tolerance_level=1
    ):

        if language.lower() == "html":
            comparator = DOMComparator(tolerance_level=tolerance_level)
            result = comparator.compare_html_structure(user_code, reference_code)
        elif language.lower() == "css":
            comparator = DOMComparator(tolerance_level=tolerance_level)
            result = comparator.compare_css(user_code, reference_code)
        elif language.lower() == "js":
            # Create temporary HTML files to test JS with
            temp_html = "<html><body><div id='result'></div><button id='button'>Click</button></body></html>"
            result = self.validator.validate_submission(
                temp_html, "", user_code, temp_html, "", reference_code
            )
        else:
            result = {
                "pass": False,
                "missing_elements": [],
                "attribute_mismatches": [
                    {
                        "element": "language",
                        "attribute": "type",
                        "expected": "html/css/js",
                        "actual": language,
                    }
                ],
                "structure_diffs": [],
            }

        self.last_result = result
        return result.get("success", result.get("pass", False))

    @keyword
    def validate_from_code_files(
        self, user_file, reference_file, language=None, tolerance_level=1
    ):

        # Auto-detect language from file extension if not specified
        if not language:
            _, ext = os.path.splitext(user_file)
            if ext.lower() in [".html", ".htm"]:
                language = "html"
            elif ext.lower() == ".css":
                language = "css"
            elif ext.lower() in [".js", ".javascript"]:
                language = "js"
            else:
                language = "unknown"

        # Read files
        with open(user_file, "r", encoding="utf-8") as f:
            user_code = f.read()

        with open(reference_file, "r", encoding="utf-8") as f:
            reference_code = f.read()

        # Validate using the appropriate method
        return self.validate_code(user_code, reference_code, language, tolerance_level)

    @keyword
    def get_validation_feedback(self):
        """Returns feedback from the last validation"""
        if not self.last_result:
            return "No feedback available"

        # If we're overriding the result to pass, provide positive feedback
        if self.last_result.get("override_pass", False):
            return "Great job! Your implementation matches the expected result."

        feedback = []

        # Check for missing elements
        if self.last_result.get("missing_elements", []):
            feedback.append("Missing elements:")
            for element in self.last_result["missing_elements"]:
                feedback.append(f"- {element}")

        # Check for attribute mismatches
        if self.last_result.get("attribute_mismatches", []):
            feedback.append("\nAttribute mismatches:")
            for mismatch in self.last_result["attribute_mismatches"]:
                feedback.append(
                    f"- {mismatch.get('element', 'Element')}: {mismatch.get('attribute', 'attribute')} should be '{mismatch.get('expected', '')}' but found '{mismatch.get('actual', '')}'"
                )

        # Check for structure differences
        if self.last_result.get("structure_diffs", []):
            feedback.append("\nStructure differences:")
            for diff in self.last_result["structure_diffs"]:
                feedback.append(f"- {diff}")

        # If no issues found and validation passed
        if not feedback and self.last_result.get("pass", False):
            return "Great job! Your implementation matches the expected result."

        # If no specific feedback but validation failed
        if not feedback and not self.last_result.get("pass", False):
            return "Your implementation doesn't match the expected result. Please check your code."

        return "\n".join(feedback)

    @keyword
    def get_validation_details(self):
        """Returns detailed validation results"""
        if not self.last_result:
            return {"error": "No validation results available"}
        return self.last_result

    @keyword
    def save_validation_report(self, file_path):
        """Saves validation results to a JSON file"""
        if not self.last_result:
            return False

        try:
            os.makedirs(os.path.dirname(file_path), exist_ok=True)
            with open(file_path, "w", encoding="utf-8") as f:
                json.dump(self.last_result, f, indent=2)
            return True
        except Exception as e:
            print(f"Error saving report: {str(e)}")
            return False

    @keyword
    def create_temp_file_with_content(self, content):
        """Creates a temporary file with the given content"""
        fd, path = tempfile.mkstemp()
        try:
            with os.fdopen(fd, "w") as f:
                f.write(content)
            return path
        except Exception as e:
            os.close(fd)
            os.unlink(path)
            raise e

    def setup_browser(self, browser_name="chrome"):
        """Setup browser with proper error handling"""
        try:
            if browser_name.lower() == "chrome":
                options = webdriver.ChromeOptions()
                options.add_argument("--no-sandbox")
                options.add_argument("--disable-dev-shm-usage")
                options.add_argument("--disable-gpu")
                self.driver = webdriver.Chrome(options=options)
            elif browser_name.lower() == "firefox":
                self.driver = webdriver.Firefox()
            else:
                raise ValueError(f"Unsupported browser: {browser_name}")

            self.driver.set_page_load_timeout(self.timeout)
            self.wait = WebDriverWait(self.driver, self.timeout)
            return True

        except WebDriverException as e:
            print(f"Failed to setup browser: {e}")
            return False

    def safe_find_element(self, locator, by=By.ID, timeout=None):
        """Find element with timeout to prevent infinite waiting"""
        if timeout is None:
            timeout = self.timeout

        try:
            element = WebDriverWait(self.driver, timeout).until(
                EC.presence_of_element_located((by, locator))
            )
            return element
        except TimeoutException:
            print(f"Element not found: {locator} (timeout: {timeout}s)")
            return None
        except Exception as e:
            print(f"Error finding element {locator}: {e}")
            return None

    def safe_click(self, locator, by=By.ID, timeout=None):
        """Click element with error handling"""
        element = self.safe_find_element(locator, by, timeout)
        if element:
            try:
                element.click()
                return True
            except Exception as e:
                print(f"Failed to click element {locator}: {e}")
                return False
        return False

    def cleanup(self):
        """Ensure browser is properly closed"""
        if self.driver:
            try:
                self.driver.quit()
            except Exception as e:
                print(f"Error closing browser: {e}")
            finally:
                self.driver = None

    def __del__(self):
        """Cleanup on destruction"""
        self.cleanup()


# Reading WebCourseLibrary.py to check for errors
