NewFormatOfRobot/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asgiref==3.8.1
astroid==3.3.5
asttokens==3.0.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
backcall==0.2.0
backports.tarfile==1.2.0
bcrypt==4.0.1
beautifulsoup4>=4.11
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
BTrees==6.1
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
constantly==23.10.4
contourpy==1.2.1
crypto==1.4.1
cryptography==44.0.1
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
Django==5.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.6.1
docopt==0.6.2
docutils==0.21.2
drf-nested-routers==0.94.1
email_validator==2.2.0
executing==2.2.0
fanstatic==1.4
fastapi>=0.95
fastapi-cli==0.0.5
fastjsonschema==2.21.1
filelock==3.14.0
Flask>=2.2
fonttools==4.52.4
fsspec==2024.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.1
httpx==0.28.1
hyperlink==21.0.0
id==1.5.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.0.0
ipython==8.12.3
iqoptionapi @ git+https://github.com/Lu-Yi-Hsun/iqoptionapi.git@e2a198fe74a3cd646604b11d293c383ef3d39330
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jose==1.0.0
js==1.0
jsonpath-ng==1.7.0
jsonschema>=4.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.6.0
kiwisolver==1.4.5
lazy_loader==0.4
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.6.0
mpmath==1.3.0
Naked==0.1.32
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
nh3==0.2.20
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
persistent==6.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
prompt_toolkit==3.0.50
psutil>=5.9
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.0
pylint==3.3.2
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.16.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2024.1
PyYAML>=6.0
pyzmq==26.2.1
readme_renderer==44.0
redis==5.2.1
referencing==0.36.2
reflex==0.6.8
reflex-chakra==0.6.3
reflex-hosting-cli==0.1.35
requests>=2.28
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
robotframework
robotframework-jsonlibrary==0.5
robotframework-requests
rpds-py==0.23.0
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
service-identity==24.2.0
shellescape==3.8.1
shellingham==1.5.4
shiboken6==6.8.0.2
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlmodel==0.0.22
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
starlette-admin==0.14.1
sympy==1.12.1
tabulate==0.9.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
tomlkit==0.13.2
torch==2.3.0
torchvision==0.18.0
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
transaction==5.0
twine==6.1.0
Twisted==24.11.0
txaio==23.1.1
typer==0.15.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.26
urllib3>=1.26
utils==1.0.2
uvicorn>=0.20
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.9
websocket-client==0.56.0
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarg==0.1.9
zc.lockfile==3.0.post1
ZConfig==4.1
zipp==3.21.0
ZODB==6.0
zodbpickle==4.1.1
zope.deferredimport==5.0
zope.interface==7.1.0
zope.proxy==6.1
